Migrate from Next.js to Astro
Convert the project from Next.js Pages Router to Astro with React
islands. The interactive podcast list and contact form are kept as
React components hydrated client-side; the layout shell, page wrapper,
and floating wide-monitor widget render as Astro/HTML.

- Replace next + eslint-config-next with astro, @astrojs/react,
  @astrojs/vercel, and @tailwindcss/vite
- Move source under src/ following Astro conventions
- Convert pages/api/podcast-suggestion.ts to an Astro APIRoute, using
  the Web Request/Response API and import.meta.env
- Build-time iTunes enrichment moves from getStaticProps into the
  index.astro frontmatter
- Rename NEXT_PUBLIC_TURNSTILE_SITE_KEY to PUBLIC_TURNSTILE_SITE_KEY
