+++ .env.example
@@ -1,5 +1,5 @@
 TELEGRAM_BOT_TOKEN=
 TELEGRAM_CHAT_ID=
 
-NEXT_PUBLIC_TURNSTILE_SITE_KEY="1x00000000000000000000AA"
+PUBLIC_TURNSTILE_SITE_KEY="1x00000000000000000000AA"
 TURNSTILE_SECRET="1x0000000000000000000000000000000AA"
