+++ .gitignore
@@ -7,3 +7,7 @@ out
 .now
 .vercel
 tsconfig.tsbuildinfo
+
+# Astro
+dist
+.astro
