+++ astro.config.mjs
@@ -1,0 +1,13 @@
+import { defineConfig } from "astro/config";
+import react from "@astrojs/react";
+import vercel from "@astrojs/vercel";
+import tailwindcss from "@tailwindcss/vite";
+
+export default defineConfig({
+  site: "https://podcasts.michaelbonner.dev",
+  integrations: [react()],
+  adapter: vercel(),
+  vite: {
+    plugins: [tailwindcss()],
+  },
+});
