--- eslint.config.mjs
@@ -1,5 +1,0 @@
-import nextConfig from "eslint-config-next/core-web-vitals";
-
-const config = [...nextConfig];
-
-export default config;
