--- next.config.js
@@ -1,10 +1,0 @@
-module.exports = {
-  images: {
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "**.mzstatic.com",
-      },
-    ],
-  },
-};
