+++ package.json
@@ -2,29 +2,27 @@
   "name": "podcasts-i-listen-to",
   "version": "1.0.0",
   "description": "Runs here: [https://podcasts.michaelbonner.dev/](https://podcasts.michaelbonner.dev/)",
-  "main": "index.js",
+  "type": "module",
   "dependencies": {
+    "@astrojs/react": "^4.4.0",
+    "@astrojs/vercel": "^8.2.7",
     "@marsidev/react-turnstile": "^1.5.2",
-    "next": "^16.2.4",
+    "astro": "^5.15.0",
     "react": "^19.2.5",
     "react-dom": "^19.2.5"
   },
   "devDependencies": {
-    "@tailwindcss/postcss": "^4.2.4",
+    "@tailwindcss/vite": "^4.2.4",
     "@types/react": "^19.2.14",
-    "eslint": "^9.39.4",
-    "eslint-config-next": "^16.2.4",
-    "postcss": "^8.5.14",
+    "@types/react-dom": "^19.2.3",
     "tailwindcss": "^4.2.4",
     "typescript": "^6.0.3"
   },
   "scripts": {
-    "build": "next build",
-    "deploy": "npm run build && npm run export",
-    "dev": "next",
-    "export": "npm run build && next export",
-    "start": "next start",
-    "lint": "eslint .",
+    "dev": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview",
+    "astro": "astro",
     "out": "bun outdated",
     "up": "bun update"
   },
