--- pages/_app.tsx
@@ -1,8 +1,0 @@
-import { AppProps } from "next/app";
-import "../styles/index.css";
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
-}
-
-export default MyApp;
