--- pages/api/podcast-suggestion.ts
@@ -1,135 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from "next";
-
-type TurnstileResponse = {
-  success: boolean;
-};
-const verifyTurnstileToken = async (
-  token: string
-): Promise<TurnstileResponse> => {
-  const body = `secret=${encodeURIComponent(
-    process.env.TURNSTILE_SECRET
-  )}&response=${encodeURIComponent(token)}`;
-
-  const res = await fetch(
-    "https://challenges.cloudflare.com/turnstile/v0/siteverify",
-    {
-      method: "POST",
-      body,
-      headers: {
-        "content-type": "application/x-www-form-urlencoded",
-      },
-    }
-  );
-
-  const data = (await res.json()) as TurnstileResponse;
-  return data;
-};
-
-type ReturnData = {
-  success: boolean;
-  data: string;
-};
-
-const postToTelegram = async (text: string) => {
-  try {
-    const res = await fetch(
-      `https://api.telegram.org/bot${process.env.TELEGRAM_BOT_TOKEN}/sendMessage`,
-      {
-        method: "POST",
-        headers: { "content-type": "application/json" },
-        body: JSON.stringify({
-          chat_id: process.env.TELEGRAM_CHAT_ID,
-          text,
-          parse_mode: "HTML",
-        }),
-      }
-    );
-    return res.ok;
-  } catch (error) {
-    console.error("error", error);
-    return false;
-  }
-};
-
-const handler = async (
-  req: NextApiRequest,
-  res: NextApiResponse<ReturnData>
-) => {
-  res.setHeader("Content-Type", "application/json");
-
-  if (!process.env.TELEGRAM_BOT_TOKEN || !process.env.TELEGRAM_CHAT_ID) {
-    res.statusCode = 400;
-    res.end(
-      JSON.stringify({
-        success: false,
-        data: `The server is not configured correctly to accept this request. Please let me know.`,
-      })
-    );
-    return;
-  }
-
-  if (req.method !== "POST") {
-    res.statusCode = 400;
-    res.end(
-      JSON.stringify({
-        success: false,
-        data: `Must be post request. You sent ${req.method}`,
-      })
-    );
-    return;
-  }
-
-  if (!req.body) {
-    res.statusCode = 400;
-    res.end(JSON.stringify({ success: false, data: "Missing form details" }));
-    return;
-  }
-
-  if (!req.body.podcast_name) {
-    res.statusCode = 400;
-    res.end(JSON.stringify({ success: false, data: "Missing podcast name" }));
-    return;
-  }
-
-  if (!req.body.podcast_url) {
-    res.statusCode = 400;
-    res.end(JSON.stringify({ success: false, data: "Missing podcast url" }));
-    return;
-  }
-
-  if (!req.body.your_name) {
-    res.statusCode = 400;
-    res.end(JSON.stringify({ success: false, data: "Missing your name" }));
-    return;
-  }
-
-  // captcha
-  const captchaVerification = await verifyTurnstileToken(req.body.token);
-
-  // if the captcha fails, assume it's a bot and give them a fail
-  if (captchaVerification.success !== true) {
-    res.statusCode = 400;
-    return res.end(
-      JSON.stringify({ success: false, data: "Could not send notification" })
-    );
-  }
-
-  const postIt = await postToTelegram(
-    `${req.body.your_name} thinks you should check out <a href="${req.body.podcast_url}">${req.body.podcast_name}</a>`
-  );
-
-  if (postIt) {
-    res.statusCode = 200;
-    res.end(
-      JSON.stringify({ success: true, data: "Saved and sent notification" })
-    );
-  } else {
-    console.log("postIt", postIt);
-    res.statusCode = 400;
-    res.end(
-      JSON.stringify({ success: false, data: "Could not send notification" })
-    );
-  }
-  return;
-};
-export default handler;
